#14 Packages refactor
