+++ app/src/main/java/com/example/triviaapp/common/constants/DataStoreConstants.kt
@@ -1,4 +1,4 @@
-package com.example.triviaapp.common
+package com.example.triviaapp.common.constants
 
 object DataStoreConstants {
     const val DATA_STORE_PREFERENCES_NAME_FOR_STATS = "stats"
