+++ app/src/main/java/com/example/triviaapp/common/constants/NetworkConstants.kt
@@ -1,7 +1,6 @@
-package com.example.triviaapp.common
+package com.example.triviaapp.common.constants
 
 object NetworkConstants {
 
     const val BASE_URL = "https://kmaslowiec.github.io/jsonRepo/"
-
 }
