+++ app/src/main/java/com/example/triviaapp/game/model/Questions.kt
@@ -1,3 +1,3 @@
-package com.example.triviaapp.model
+package com.example.triviaapp.game.model
 
 class Questions : ArrayList<QuestionsItem>()
