+++ app/src/main/java/com/example/triviaapp/navigation/AppNavigation.kt
@@ -7,7 +7,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.triviaapp.homescreen.HomeScreen
+import com.example.triviaapp.game.screen.HomeScreen
 
 @Composable
 fun AppNavigation(modifier: Modifier) {
