+++ app/src/test/java/com/example/triviaapp/repository/game/repository/GameRepositoryTest.kt
@@ -1,7 +1,8 @@
-package com.example.triviaapp.repository
+package com.example.triviaapp.repository.game.repository
 
-import com.example.triviaapp.model.Questions
-import com.example.triviaapp.service.QuestionService
+import com.example.triviaapp.game.model.Questions
+import com.example.triviaapp.game.repository.GameRepository
+import com.example.triviaapp.game.api.GameApi
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
@@ -13,21 +14,21 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.IOException
 
-class QuestionRepositoryTest {
+class GameRepositoryTest {
 
-    private lateinit var questionService: QuestionService
-    private lateinit var questionRepository: QuestionRepository
+    private lateinit var gameApi: GameApi
+    private lateinit var questionRepository: GameRepository
 
     @BeforeEach
     fun setup() {
-        questionService = mockk()
-        questionRepository = QuestionRepository(questionService)
+        gameApi = mockk()
+        questionRepository = GameRepository(gameApi)
     }
 
     @Test
     fun `getQuestions returns data on successful API call`() = runBlocking {
         val expectedQuestions = Questions()
-        coEvery { questionService.getAllQuestions() } returns expectedQuestions
+        coEvery { gameApi.getAllQuestions() } returns expectedQuestions
 
         val result = questionRepository.getQuestions()
 
@@ -39,7 +40,7 @@ class QuestionRepositoryTest {
     @Test
     fun `getQuestions returns exception on API call failure`() = runBlocking {
         val exception = IOException("Network Error")
-        coEvery { questionService.getAllQuestions() } throws exception
+        coEvery { gameApi.getAllQuestions() } throws exception
 
         val result = questionRepository.getQuestions()
